Ignore cert errors
This fixes my initial issue from #34. It does not add the extra jargle in that issue.

It should also be noted that with this change you should never try to hit any site you wish to be secure on. Such as you should not log into Amazon with it. I don't think that would be an issue. But people are unwise sometimes.