Specify host when getting remote debugging port
By default, Chrome only binds to `127.0.0.1` when specifying a remote
debugging port, but `get-port` (as a result of the default behavior of
node.js) checks on the unspecified IPv6 or IPv4 address when determining
if a port is in use or not.

Since they are different addresses, `get-port` won't detect another
Chrome or node.js process that has already been bound to port 9222. This
updates the usage of `get-port` to specify `127.0.0.1` when getting a
remote debugging port.