Improve attach debug port resolution
Attach debug types shouldn't need to explicitly specifcy a debug port
since, if a Browser Preview window is open, the window manager should be
able to provide the debug port. If it can't, that means a Browser
Preview window isn't open. If that's the case, display an error message
with guidance to open a Browser Preview window or use the "launch" debug
type instead.